Started setting up messagelistener commands to mimic existing capability
