+++ src/main/java/org/dylansbogar/MessageListener.java
@@ -1,0 +1,35 @@
+package org.dylansbogar;
+
+import net.dv8tion.jda.api.entities.Message;
+import net.dv8tion.jda.api.entities.channel.middleman.MessageChannel;
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
+import net.dv8tion.jda.api.hooks.ListenerAdapter;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class MessageListener extends ListenerAdapter {
+    private static final String ozbargain = "https://www.ozbargain.com.au/node/";
+
+    Pattern thanksPattern = Pattern.compile("thanks man", Pattern.CASE_INSENSITIVE);
+    @Override
+    public void onMessageReceived(MessageReceivedEvent event)
+    {
+        // Ensure griddybot does not respond to itself.
+        if (event.getAuthor().isBot()) return;
+
+        Message message = event.getMessage();
+        String content = message.getContentRaw();
+        MessageChannel channel = event.getChannel();
+
+        Matcher thanks = thanksPattern.matcher(content);
+
+        if (content.startsWith(ozbargain)) {
+            channel.sendMessage("Thanks just bought").queue();
+        } else if (content.equalsIgnoreCase("gm") || content.equalsIgnoreCase("gn")) {
+            channel.sendMessage(content).queue();
+        } else if (thanks.find()) {
+            channel.sendMessage("No worries <3").queue();
+        }
+    }
+}
